refactor(attempts): 시도할 횟수 입력받는 과정을 racingGameController에서 분리
