+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import racingcar.attempts.AttemptsController;
 import racingcar.car.CarController;
 import racingcar.game.RacingGameController;
 
@@ -7,7 +8,9 @@ public class Application {
 	public static void main(String[] args) {
 		// TODO 구현 진행
 		CarController carController = new CarController();
-		RacingGameController racingGameController = new RacingGameController(carController.createCars());
+		AttemptsController attemptsController = new AttemptsController();
+		RacingGameController racingGameController = new RacingGameController(carController.createCars(),
+			attemptsController.selectAttempts());
 		racingGameController.run();
 	}
 }
