+++ src/main/java/racingcar/game/RacingGameController.java
@@ -1,38 +1,22 @@
 package racingcar.game;
 
-import racingcar.view.ErrorView;
-import racingcar.view.InputView;
+import racingcar.attempts.Attempts;
 import racingcar.view.OutputView;
 
 public class RacingGameController {
 	private RacingGame racingGame;
 	private Attempts attempts;
 
-	public RacingGameController(RacingGame racingGame) {
+	public RacingGameController(RacingGame racingGame, Attempts attempts) {
 		this.racingGame = racingGame;
+		this.attempts = attempts;
 	}
 
 	public void run() {
-		requestAttempts();
 		startRacing();
 		announceWinners();
 	}
-
-	private void requestAttempts() {
-		attempts = new Attempts();
-		InputView.showAttemptsMessage();
-		inputAttempts(attempts);
-	}
-
-	private void inputAttempts(Attempts attempts) {
-		try {
-			attempts.input();
-		} catch (IllegalArgumentException illegalArgumentException) {
-			ErrorView.show(illegalArgumentException.getMessage());
-			inputAttempts(attempts);
-		}
-	}
-
+	
 	private void startRacing() {
 		OutputView.showResultMessage();
 		do {
